make it easy to remove rules from the "recommended" config
Summary:

Allow setting `defaultLevel: 'off'` to exclude a rule from the generated `recommended` config.
This makes it easier to add selectors which should only be turned on in  a subset of the codebase because you no longer need to disable it in the base config.

Test Plan:

```
pnpm run ci
```